+++ src/js/tracking-you/index.js
@@ -1,0 +1,40 @@
+export default function applyAccordionListeners() {
+  console.log('applying listeners')
+  document.querySelectorAll('cwds-accordion').forEach((acc) => {
+    console.log(acc)
+    acc.addEventListener('click',function() {
+      console.log('click')
+      if(this.querySelector('.accordion-title')) {
+        console.log(this)
+        reportGA('accordion', this.querySelector('.accordion-title').textContent.trim())
+      }
+    });
+  });
+
+  function reportGA(elementType,eventString) {
+    if(typeof(gtag) !== 'undefined') {
+      gtag('event','click',{'event_category':elementType,'event_label':eventString});
+    } else {
+      setTimeout(function() {
+        reportGA(elementType,eventString)
+      }, 500);
+    }
+  }
+
+  document.querySelectorAll('.show-all').forEach(btn => {
+    const isCloseButton = btn.classList.contains('d-none');
+    btn.addEventListener('click', () => {
+      let wait = 0;
+      document
+        .querySelectorAll(
+          `cwds-step-list .list-group-item-action${
+            isCloseButton ? '.list-open' : ':not(.list-open)'
+          }`
+        )
+        .forEach(lstitem => setTimeout(() => lstitem.click(), 50 * wait++));
+      document
+        .querySelectorAll('.show-all')
+        .forEach(y => y.classList.toggle('d-none'));
+    });
+  });
+}
